+++ src/EventHub/EventHub/Cmdlets/AuthorizationRule/SetAzureEventhubAuthorizationRules.cs
@@ -94,7 +94,7 @@ namespace Microsoft.Azure.Commands.EventHub.Commands
                     {
                         if (ShouldProcess(target: sasRule.Name, action: string.Format(Resources.UpdateEventHubAuthorizationrule, Name, EventHub)))
                         {
-                            WriteObject(Client.CreateOrUpdateEventHubAuthorizationRules(ResourceGroupName, Namespace, EventHub, Name, sasRule));
+                            WriteObject(UtilityClient.CreateOrUpdateEventHubAuthorizationRules(ResourceGroupName, Namespace, EventHub, Name, sasRule));
                         }
                     }
                     else
@@ -102,7 +102,7 @@ namespace Microsoft.Azure.Commands.EventHub.Commands
                         if (ShouldProcess(target: sasRule.Name, action: string.Format(Resources.UpdateNamespaceAuthorizationrule, Name, Namespace)))
                         {
                             sasRule = InputObject;
-                            WriteObject(Client.CreateOrUpdateNamespaceAuthorizationRules(ResourceGroupName, Namespace, Name, sasRule));
+                            WriteObject(UtilityClient.CreateOrUpdateNamespaceAuthorizationRules(ResourceGroupName, Namespace, Name, sasRule));
                         }
                     }
 
@@ -110,7 +110,7 @@ namespace Microsoft.Azure.Commands.EventHub.Commands
                 if (ParameterSetName.Equals(NamespaceAuthoRuleParameterSet))
                     if (ShouldProcess(target: sasRule.Name, action: string.Format(Resources.UpdateNamespaceAuthorizationrule, Name, Namespace)))
                     {
-                        WriteObject(Client.CreateOrUpdateNamespaceAuthorizationRules(ResourceGroupName, Namespace, Name, sasRule));
+                        WriteObject(UtilityClient.CreateOrUpdateNamespaceAuthorizationRules(ResourceGroupName, Namespace, Name, sasRule));
                     }
 
 
@@ -118,7 +118,7 @@ namespace Microsoft.Azure.Commands.EventHub.Commands
                 if (ParameterSetName.Equals(EventhubAuthoRuleParameterSet))
                     if (ShouldProcess(target: sasRule.Name, action: string.Format(Resources.UpdateEventHubAuthorizationrule, Name, EventHub)))
                     {
-                        WriteObject(Client.CreateOrUpdateEventHubAuthorizationRules(ResourceGroupName, Namespace, EventHub, Name, sasRule));
+                        WriteObject(UtilityClient.CreateOrUpdateEventHubAuthorizationRules(ResourceGroupName, Namespace, EventHub, Name, sasRule));
                     }
             }
             catch (Management.EventHub.Models.ErrorResponseException ex)