+++ tools/VersionController/Program.cs
@@ -28,11 +28,12 @@ namespace VersionController
     public class Program
     {
         private static VersionBumper _versionBumper;
-        private static VersionValidator _versionValidator;
 
         private static Dictionary<string, AzurePSVersion> _minimalVersion = new Dictionary<string, AzurePSVersion>();
         private static List<string> _projectDirectories, _outputDirectories;
         private static string _rootDirectory, _moduleNameFilter;
+
+        private const string Psd1NameExtension = ".psd1";
 
         private static IList<string> ExceptionFileNames = new List<string>()
         {
@@ -70,13 +71,12 @@ namespace VersionController
             _moduleNameFilter = string.Empty;
             if (args != null && args.Length > 1)
             {
-                _moduleNameFilter = args[1] + ".psd1";
+                _moduleNameFilter = args[1] + Psd1NameExtension;
             }
 
             ConsolidateExceptionFiles(exceptionsDirectory);
             ValidateManifest();
             BumpVersions();
-            ValidateVersionBump();
         }
 
         private static void ValidateManifest()
@@ -113,6 +113,7 @@ namespace VersionController
         {
             return File.Exists(Path.Combine(directory, "readme.md"));
         }
+
         /// <summary>
         /// Bump the version of changed modules or a specified module.
         /// </summary>
@@ -175,6 +176,10 @@ namespace VersionController
                 var file = File.ReadAllLines(miniVersionFile);
                 var header = file.First();
                 var lines = file.Skip(1).Where(c => !string.IsNullOrEmpty(c));
+
+                var bumpingModule = _moduleNameFilter.Replace(Psd1NameExtension, "");
+                List<string> _minimalVersionContent = new List<string>() { header };
+
                 foreach (var line in lines)
                 {
                     var cols = line.Split(",").Select(c => c.StartsWith("\"") ? c.Substring(1) : c)
@@ -183,11 +188,17 @@ namespace VersionController
                     if (cols.Length >= 2)
                     {
                         _minimalVersion.Add(cols[0], new AzurePSVersion(cols[1]));
+
+                        // Bump one module, only remove its minimal version from MinimalVersion.csv content
+                        if (!string.IsNullOrEmpty(bumpingModule) && !cols[0].Equals(bumpingModule))
+                        {
+                            _minimalVersionContent.Add(line);
+                        }
                     }
                 }
 
                 // Clean MinimalVersion.csv
-                File.WriteAllLines(Path.Combine(_rootDirectory, @"tools\VersionController", "MinimalVersion.csv"), new string[]{ header});
+                File.WriteAllLines(Path.Combine(_rootDirectory, @"tools\VersionController", "MinimalVersion.csv"), _minimalVersionContent.ToArray());
             }
 
             //Make Az.Accounts as the last module to calculate
@@ -222,51 +233,7 @@ namespace VersionController
                 _versionBumper.BumpAllVersions();
             }
         }
-
-        /// <summary>
-        /// Validate version bump of changed modules or a specified module.
-        /// </summary>
-        private static void ValidateVersionBump()
-        {
-            var changedModules = new List<string>();
-            foreach (var directory in _projectDirectories)
-            {
-                var changeLogs = Directory.GetFiles(directory, "ChangeLog.md", SearchOption.AllDirectories)
-                                            .Where(f => !ModuleFilter.IsAzureStackModule(f))
-                                            .Select(f => GetModuleManifestPath(Directory.GetParent(f).FullName))
-                                            .Where(m => !string.IsNullOrEmpty(m) && m.Contains(_moduleNameFilter))
-                                            .ToList();
-                changedModules.AddRange(changeLogs);
-            }
-
-            foreach (var projectModuleManifestPath in changedModules)
-            {
-                var moduleFileName = Path.GetFileName(projectModuleManifestPath);
-                var moduleName = moduleFileName.Replace(".psd1", "");
-
-                var outputModuleManifest = _outputDirectories
-                                            .SelectMany(d => Directory.GetDirectories(d, moduleName))
-                                            .SelectMany(d => Directory.GetFiles(d, moduleFileName))
-                                            .ToList();
-                if (outputModuleManifest.Count == 0)
-                {
-                    throw new FileNotFoundException("No module manifest file found in output directories");
-                }
-                else if (outputModuleManifest.Count > 1)
-                {
-                    throw new IndexOutOfRangeException("Multiple module manifest files found in output directories");
-                }
-
-                var outputModuleManifestFile = outputModuleManifest.FirstOrDefault();
-
-                var validatorFileHelper = new VersionFileHelper(_rootDirectory, outputModuleManifestFile, projectModuleManifestPath);
-
-                _versionValidator = new VersionValidator(validatorFileHelper);
-
-                _versionValidator.ValidateAllVersionBumps();
-            }
-        }
-
+        
         /// <summary>
         /// Check if a change log has anything under the Upcoming Release header.
         /// </summary>